normalized json decoding in jsondecoder.go
