+++ CITATION.cff
@@ -5,6 +5,6 @@ authors:
    - family-names: Doiel
      given-names: R. S.
      orcid: https://orcid.org/0000-0003-0900-6903
-version: 1.1.0-rc2
+version: 1.1.0-rc3
 title: eprinttools
 date-released: 2021-11-18
