+++ crossref_test.go
@@ -1,23 +1,9 @@
 package eprinttools
 
 import (
-	"bytes"
-	"encoding/json"
-	"io"
 	"io/ioutil"
 	"testing"
 )
-
-// Custom JSON decoder so we can treat numbers easier
-func jsonDecode(src []byte, obj interface{}) error {
-	dec := json.NewDecoder(bytes.NewReader(src))
-	dec.UseNumber()
-	err := dec.Decode(&obj)
-	if err != nil && err != io.EOF {
-		return err
-	}
-	return nil
-}
 
 func TestIssue35(t *testing.T) {
 	doi := `10.1093/mnras/stab2505`
