+++ ep3apid_test.go
@@ -2,7 +2,6 @@ package eprinttools
 
 import (
 	"bytes"
-	"encoding/json"
 	"fmt"
 	"io/ioutil"
 	"net/http"
@@ -102,7 +101,7 @@ func runWriteTest(t *testing.T, api *EP3API, repoID string, repo *DataSource, ro
 		}
 		t.Logf(`Post returned: %s`, src)
 		ids := []int{}
-		if err := json.Unmarshal(src, &ids); err != nil {
+		if err := jsonDecode(src, &ids); err != nil {
 			t.Errorf(`Failed to unmarshal post results fo %q, %s`, u, err)
 			t.FailNow()
 		}
@@ -123,7 +122,7 @@ func runReadTests(t *testing.T, api *EP3API, repoID string, route string) {
 			t.FailNow()
 		}
 		repository := map[string][]string{}
-		if err := json.Unmarshal(src, &repository); err != nil {
+		if err := jsonDecode(src, &repository); err != nil {
 			t.Errorf(`Failed unmarshal %s, %s`, u, err)
 			t.FailNow()
 		}
@@ -134,7 +133,7 @@ func runReadTests(t *testing.T, api *EP3API, repoID string, route string) {
 			t.FailNow()
 		}
 		keys := []int{}
-		if err := json.Unmarshal(src, &keys); err != nil {
+		if err := jsonDecode(src, &keys); err != nil {
 			t.Errorf(`Failed %s, %s`, u, err)
 			t.FailNow()
 		}
