+++ version.go
@@ -1,4 +1,4 @@
 package eprinttools
 
-const Version = "1.1.0-rc2"
+const Version = "1.1.0-rc3"
